apiserver/requirements.txt
django==2.1.2
djangorestframework==3.8.2
markdown==3.0.1
django-filter==2.0.0
redis==2.10.6
manuale==1.1.0
requests==2.19.1
mysqlclient==1.3.13
pyopenssl==18.0.0
